refactor: move schema to model directory
