+++ src/pages/RegistryPage/model/schema.ts
