+++ src/pages/RegistryPage/ui/types.ts
@@ -1,4 +1,4 @@
-import type { schema } from '@pages/RegistryPage/ui/schema.ts';
+import type { schema } from '@pages/RegistryPage/model/schema.ts';
 import { z } from 'zod';
 
 export type RegistryForm = z.infer<typeof schema>;
